Fix UI scaling precision in KoolScreen
Changed integer division to floating-point division for UI size calculations in KoolScreen to ensure correct scaling. Also updated modVersion to 2.3.2b in gradle.properties.

+++ src/main/java/ru/hollowhorizon/hc/client/kool/KoolScreen.kt
@@ -28,8 +28,8 @@ open class KoolScreen : Screen("".literal), HudHideable {
         }
         super.init()
         uiSize?.let {
-            val w = KoolManager.context.windowWidth / it.x
-            val h = KoolManager.context.windowHeight / it.y
+            val w = KoolManager.context.windowWidth / it.x.toFloat()
+            val h = KoolManager.context.windowHeight / it.y.toFloat()
             UiScale.uiScale.set(min(w, h) / UiScale.windowScale.value)
         }
     }
